requirements.txt
pygame==2.3.0
numpy==1.24.2
pillow==9.4.0
